restaurant_service/requirements.txt
# python=3.11.3

fastapi==0.103.1
pydantic-settings==2.0.3
uvicorn[standard]==0.23.2

# SQL ORM and migrations
sqlalchemy==2.0.20
alembic==1.12.0
aiosqlite==0.19.0

httpx==0.25.0

# Celery
celery==5.3.4
flower==2.0.1
Redis==5.0.1
